Added docs about webpack config
This commit provides a webpack config for a bundle similiar to browserify.

Browserify bundles to an empty function for `fs`, and for each module declares
`fs` as a `graceful-fs` require.

https://benclinkinbeard.com/posts/how-browserify-works/

The webpack bundle built from this commit's config resolves `graceful-fs`
as an alias to `fs`.

`fs` used in dat-sdk/index for browser/hyperdrive IO:

https://github.com/random-access-storage/random-access-web